feat(pool): remove pool stat logics
